requirements-tests.txt
mock>=1.3.0,<2.0
unittest2>=1.1.0,<2.0
nose>=1.3.7
ruamel.yaml>=0.15.0
jinja2
